+++ src/vm.rs
@@ -2245,9 +2245,11 @@ fn interpret_instructions(
                         let fields = definitions
                             .iter()
                             .map(|field| {
-                                if module.type_definition(&field.ty).is_some_and(|definition| {
-                                    definition.representation == Representation::Delegate
-                                }) {
+                                if field.ty == Type::RuntimeTypeHandle
+                                    || module.type_definition(&field.ty).is_some_and(|definition| {
+                                        definition.representation == Representation::Delegate
+                                    })
+                                {
                                     Ok(Value::Uninitialized(field.ty.clone()))
                                 } else {
                                     crate::initialization::default_value(module, &field.ty)